refactor: abstract chatheader and chatfooter component
